Subnet IP Address and Host Calculator
This program takes an IP address and the desired number of subnets as input. It determines the IP class, calculates the subnet mask, and converts it to binary. The program adjusts the binary subnet address based on the number of borrowed bits, then calculates and displays the binary subnet IP address and the number of possible hosts per subnet.